requirements.txt
Flask==2.0.2
Click==8.0.3
flask-restful==0.3.9
tinydb==4.5.2
bcrypt-flask==1.0.2
flask-jwt-extended==4.3.1
pyfiglet==0.8.post1
psutil==5.8.0
libvirt-python==7.9.0
lxml==4.6.4